Moves to a standalone core file instead of staying in helpers.
Reuses the existing stack if a newly requested one is smaller and deallocates/creates a new stack buffer is a larger size is requested

Adds a note that this should only be used for calling outside the task. Self deleting tasks cause problems on multicore devices with rapid starts/stops
